pyproject.toml
[project]
name = "nightcharge"
version = "0.1.0"
description = "Stop your ChargePoint charging session on a schedule — the thing ChargePoint won't let you do."
authors = [
    { name = "Sriharsha Bhavani", email = "sbhavani@users.noreply.github.com" },
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.9",
    "pydantic>=2.12",
    "click>=8.1",
    "tomli>=2.0; python_version < '3.11'",
]

[project.scripts]
nightcharge = "nightcharge.__main__:cli"

[project.optional-dependencies]
dev = [
    "tox>=3.24.5",
    "black>=25.1.0",
    "pytest>=7.0",
    "aioresponses>=0.7",
    "pytest-asyncio>=0.23",
    "flake8>=7.2.0",
    "pytest-mock>=3.7.0",
    "pytest-cov>=3.0.0",
    "mypy>=1.19.1",
    "pyright>=1.1.408",
    "pre-commit>=4.5.1",
    "python-dotenv>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["nightcharge"]

[tool.mypy]
plugins = ["pydantic.mypy"]
ignore_missing_imports = true
check_untyped_defs = true
